boss file isolation
