+++ config/splat.us.stnp3.yaml
@@ -59,6 +59,8 @@ segments:
       - [0x320DC, rodata]
       - [0x32360, rodata]
       - [0x3246C, c]
+      - [0x36990, c] # Gaibon & Slogra boss
+      - [0x390BC, c]
       - [0x3DEF4, c]
       - [0x490E8, c] # EntityBoneScimitar
       - [0x4997C, c]
