+++ config/symbols.us.stnp3.txt
@@ -26,6 +26,7 @@ EntityFallingRock = 0x801B5C5C;
 EntitySwitch = 0x801B5F60;
 EntityUnkId49 = 0x801B644C;
 EntitySlogra = 0x801B6990;
+EntitySlograSpear = 0x801B75EC;
 EntitySlograSpearProjectile = 0x801B7798;
 EntityGaibon = 0x801B78BC;
 EntitySmallGaibonProjectile = 0x801B8D84;
