+++ src/st/np3/3246C.c
@@ -999,1104 +999,3 @@ void func_801B653C(void) {
 }
 
 INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801B65FC);
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", EntitySlogra);
-
-void func_801B75EC(Entity* self) {
-    s8* hitbox;
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180B50);
-
-    case 1:
-        self->facing = self[-1].facing;
-        self->posX.i.hi = self[-1].posX.i.hi;
-        self->posY.i.hi = self[-1].posY.i.hi;
-        hitbox = D_80181454;
-        hitbox += 4 * D_8018148C[self[-1].animCurFrame];
-        self->unk10 = *hitbox++;
-        self->unk12 = *hitbox++;
-        self->hitboxWidth = *hitbox++;
-        self->hitboxHeight = *hitbox++;
-        if (self[-1].ext.generic.unk84.U8.unk2 != 0) {
-            self->step++;
-        }
-        break;
-
-    case 2:
-        switch (self->step_s) {
-        case 0:
-            self->unk19 = 4;
-            self->unk3C = 0;
-            if (self->facing != 0) {
-                self->accelerationX = -0x24000;
-            } else {
-                self->accelerationX = 0x24000;
-            }
-            self->accelerationY = -0x40000;
-            self->animCurFrame = 0x23;
-            self->flags |= FLAG_DESTROY_IF_OUT_OF_CAMERA;
-            self->step_s++;
-
-        case 1:
-            MoveEntity();
-            self->accelerationY += 0x2800;
-            self->unk1E += 0x80;
-            if (!(self->unk1E & 0xFFF)) {
-                func_801C2598(0x625);
-            }
-        }
-    }
-}
-
-// projectile fired from slogra's spear
-void EntitySlograSpearProjectile(Entity* self) {
-    Entity* entity;
-
-    if (self->flags & 0x100) {
-        entity = AllocEntity(D_8007D858, &D_8007D858[32]);
-        if (entity != NULL) {
-            CreateEntityFromEntity(ENTITY_EXPLOSION, self, entity);
-            entity->subId = 1;
-        }
-        DestroyEntity(self);
-        return;
-    }
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180B5C);
-        if (self->facing == 0) {
-            self->accelerationX = -0x40000;
-        } else {
-            self->accelerationX = 0x40000;
-        }
-
-    case 1:
-        if (AnimateEntity(D_801813D4, self) == 0) {
-            func_801BD114(2);
-        }
-        break;
-
-    case 2:
-        MoveEntity();
-        AnimateEntity(D_801813E4, self);
-        break;
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", EntityGaibon);
-
-void func_801B8CC0(Entity* self) {
-    Entity* prevEntity;
-    s16 animCurFrame;
-
-    if (self->step == 0) {
-        InitializeEntity(D_80180B68);
-        self->unk3C = 0;
-    }
-
-    prevEntity = &self[-1];
-    self->facing = prevEntity->facing;
-    self->palette = prevEntity->palette;
-    self->posX.i.hi = prevEntity->posX.i.hi;
-    self->posY.i.hi = prevEntity->posY.i.hi;
-    self->animCurFrame = 0;
-
-    if ((prevEntity->animCurFrame - 32) < 3U) {
-        self->animCurFrame = 0x26;
-    } else if (prevEntity->animCurFrame == 35) {
-        self->animCurFrame = 0x27;
-    } else if ((prevEntity->animCurFrame - 36) < 2U) {
-        self->animCurFrame = 0x28;
-    }
-
-    if (prevEntity->objectId != 0x51) {
-        DestroyEntity(self);
-    }
-}
-
-// small red projectile from gaibon
-void EntitySmallGaibonProjectile(Entity* self) {
-    if (self->flags & 0x100) {
-        self->pfnUpdate = EntityExplosion;
-        self->unk19 = 0;
-        self->step = 0;
-        self->objectId = 2;
-        self->subId = 0;
-        return;
-    }
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180B74);
-        self->animSet = 2;
-        self->animCurFrame = 1;
-        self->unk19 = 5;
-        self->unk1A = 0xC0;
-        self->accelerationX = (rcos(self->unk1E) * 0x28000) >> 0xC;
-        self->accelerationY = (rsin(self->unk1E) * 0x28000) >> 0xC;
-        self->palette = 0x81B6;
-        self->unk1E -= 0x400;
-
-    case 1:
-        MoveEntity();
-        AnimateEntity(D_801815E0, self);
-        break;
-    }
-}
-
-void EntityLargeGaibonProjectile(Entity* self) {
-    Entity* newEntity;
-
-    if (self->flags & 0x100) {
-        self->pfnUpdate = EntityExplosion;
-        self->objectId = 2;
-        self->unk19 = 0;
-        self->step = 0;
-        self->subId = 1;
-        return;
-    }
-
-    switch (self->step) {
-    case 0:
-        InitializeEntity(D_80180B80);
-        if (self->subId == 0) {
-            self->animSet = 2;
-            self->unk19 = 4;
-            self->accelerationX = (rcos(self->unk1E) * 0x38000) >> 0xC;
-            self->accelerationY = (rsin(self->unk1E) * 0x38000) >> 0xC;
-            self->palette = 0x81B6;
-            self->unk1E -= 0x400;
-        } else {
-            self->animSet = 14;
-            self->unk5A = 0x79;
-            self->unk19 = 0xD;
-            self->unk1A = 0x100;
-            self->unk6C = 0x80;
-            self->palette = 0x81F3;
-            self->blendMode = 0x30;
-            self->step = 2;
-            self->unk3C = 0;
-            self->flags |= 0x2000;
-        }
-        break;
-
-    case 1:
-        MoveEntity();
-        AnimateEntity(D_801815EC, self);
-        if (!(g_blinkTimer & 3)) {
-            newEntity = AllocEntity(D_8007D858, &D_8007D858[32]);
-            if (newEntity != NULL) {
-                CreateEntityFromEntity(0x54, self, newEntity);
-                newEntity->subId = 1;
-                newEntity->unk1E = self->unk1E;
-                newEntity->zPriority = self->zPriority + 1;
-            }
-        }
-        break;
-
-    case 2:
-        self->unk6C += 0xFE;
-        self->unk1A -= 4;
-        if (AnimateEntity(D_801815FC, self) == 0) {
-            DestroyEntity(self);
-        }
-        break;
-    }
-}
-
-s32 Random(void) {
-    g_randomNext = (g_randomNext * 0x01010101) + 1;
-    return g_randomNext >> 0x18;
-}
-
-void Update(void) {
-    s16 i;
-    Entity* entity;
-    s32* unk;
-
-    for (i = 0; i < 0x20; i++) {
-        if (D_801D33F4[i]) {
-            D_801D33F4[i]--;
-        }
-    }
-
-    unk = &D_80097410;
-    if (*unk) {
-        if (!--*unk) {
-            g_api.FreePrimitives(D_80097414);
-        }
-    }
-
-    for (entity = &g_Entities[STAGE_ENTITY_START];
-         entity < &g_Entities[TOTAL_ENTITY_COUNT]; entity++) {
-        if (!entity->pfnUpdate)
-            continue;
-
-        if (entity->step) {
-            s32 unk34 = entity->flags;
-            if (unk34 & FLAG_DESTROY_IF_OUT_OF_CAMERA) {
-                s16 posX = i = entity->posX.i.hi;
-                s16 posY = entity->posY.i.hi;
-                if (unk34 & FLAG_DESTROY_IF_BARELY_OUT_OF_CAMERA) {
-                    if ((u16)(posX + 64) > 384 || (u16)(posY + 64) > 352) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                } else {
-                    if ((u16)(posX + 128) > 512 || (u16)(posY + 128) > 480) {
-                        DestroyEntity(entity);
-                        continue;
-                    }
-                }
-            }
-
-            if ((unk34 & 0x02000000)) {
-                s16 posY = entity->posY.i.hi + g_Camera.posY.i.hi;
-                s16 test = (g_CurrentRoom.vSize * 256) + 128;
-                if (posY > test) {
-                    DestroyEntity(entity);
-                    continue;
-                }
-            }
-
-            if (unk34 & 0xF) {
-                entity->palette =
-                    D_8018164C[(entity->unk49 << 1) | (unk34 & 1)];
-                entity->flags--;
-                if ((entity->flags & 0xF) == 0) {
-                    entity->palette = entity->unk6A;
-                    entity->unk6A = 0;
-                }
-            }
-
-            if (!(unk34 & 0x20000000) || (unk34 & 0x10000000) ||
-                ((u16)(entity->posX.i.hi + 64) <= 384) &&
-                    ((u16)(entity->posY.i.hi + 64) <= 352)) {
-                if (!entity->unk58 || (entity->unk58--, unk34 & 0x100000)) {
-                    if (!D_800973FC || unk34 & 0x2100 ||
-                        (unk34 & 0x200 && !(D_8003C8C4 & 3))) {
-                        g_CurrentEntity = entity;
-                        entity->pfnUpdate(entity);
-                        entity->unk44 = 0;
-                        entity->unk48 = 0;
-                    }
-                }
-            }
-        } else {
-            g_CurrentEntity = entity;
-            entity->pfnUpdate(entity);
-            entity->unk44 = 0;
-            entity->unk48 = 0;
-        }
-    }
-}
-
-void func_801B93E8(void) {
-    Entity* entity;
-    for (entity = &g_Entities[STAGE_ENTITY_START];
-         entity < &g_Entities[TOTAL_ENTITY_COUNT]; entity++) {
-        if (!entity->pfnUpdate)
-            continue;
-
-        if (entity->step) {
-            if (!(entity->flags & FLAG_UNK_10000))
-                continue;
-            if (entity->flags & 0xF) {
-                entity->palette =
-                    D_8018164C[entity->unk49 << 1 | LOH(entity->flags) & 1];
-                entity->flags--;
-                if ((entity->flags & 0xF) == 0) {
-                    entity->palette = entity->unk6A;
-                    entity->unk6A = 0;
-                }
-            }
-        }
-
-        g_CurrentEntity = entity;
-        entity->pfnUpdate(entity);
-        entity->unk44 = 0;
-        entity->unk48 = 0;
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801B94F0);
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", EntityNumericDamage);
-
-void CreateEntityFromLayout(Entity* entity, LayoutObject* initDesc) {
-    DestroyEntity(entity);
-    entity->objectId = initDesc->objectId & 0x3FF;
-    entity->pfnUpdate = PfnEntityUpdates[entity->objectId];
-    entity->posX.i.hi = initDesc->posX - g_Camera.posX.i.hi;
-    entity->posY.i.hi = initDesc->posY - g_Camera.posY.i.hi;
-    entity->subId = initDesc->subId;
-    entity->objectRoomIndex = initDesc->objectRoomIndex >> 8;
-    entity->unk68 = (initDesc->objectId >> 0xA) & 7;
-}
-
-void CreateEntityWhenInVerticalRange(LayoutObject* layoutObj) {
-    Entity* entity;
-    s16 posY = g_Camera.posY.i.hi;
-    s16 yClose = posY - 64;
-    s16 yFar = posY + 288;
-
-    if (yClose < 0) {
-        yClose = 0;
-    }
-
-    posY = layoutObj->posY;
-    if (posY < yClose) {
-        return;
-    }
-
-    if (yFar < posY) {
-        return;
-    }
-
-    switch (layoutObj->objectId & 0xE000) {
-    case 0x0:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
-        if (entity->objectId == 0) {
-            CreateEntityFromLayout(entity, layoutObj);
-        }
-        break;
-    case 0x8000:
-        break;
-    case 0xA000:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
-        CreateEntityFromLayout(entity, layoutObj);
-        break;
-    }
-}
-
-void CreateEntityWhenInHorizontalRange(LayoutObject* layoutObj) {
-    Entity* entity;
-    s16 posX = g_Camera.posX.i.hi;
-    s16 xClose = posX - 64;
-    s16 xFar = posX + 320;
-
-    if (xClose < 0) {
-        xClose = 0;
-    }
-
-    posX = layoutObj->posX;
-    if (posX < xClose) {
-        return;
-    }
-
-    if (xFar < posX) {
-        return;
-    }
-
-    switch (layoutObj->objectId & 0xE000) {
-    case 0x0:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
-        if (entity->objectId == 0) {
-            CreateEntityFromLayout(entity, layoutObj);
-        }
-        break;
-    case 0x8000:
-        break;
-    case 0xA000:
-        entity =
-            &g_Entities[STAGE_ENTITY_START + (u8)layoutObj->objectRoomIndex];
-        CreateEntityFromLayout(entity, layoutObj);
-        break;
-    }
-}
-
-void func_801BAFA0(s16 arg0) {
-    while (true) {
-        if ((D_801D2768->posX != 0xFFFE) && (D_801D2768->posX >= (s32)arg0)) {
-            break;
-        }
-        D_801D2768++;
-    }
-}
-
-void func_801BAFEC(s16 arg0) {
-    while (true) {
-        if ((D_801D2768->posX != 0xFFFF) &&
-            ((arg0 >= D_801D2768->posX) || (D_801D2768->posX == 0xFFFE))) {
-            break;
-        }
-        D_801D2768--;
-    }
-}
-
-void func_801BB044(s16 arg0) {
-    s32 expected;
-    u8 flag;
-
-    if (D_801D2770 != 0) {
-        func_801BAFA0(arg0 - D_80097908);
-        D_801D2770 = 0;
-    }
-
-    while (true) {
-        if ((D_801D2768->posX == 0xFFFF) || (arg0 < D_801D2768->posX)) {
-            return;
-        }
-
-        expected = 0;
-        flag = (D_801D2768->objectRoomIndex >> 8) + 0xFF;
-        if ((flag == 0xFF) ||
-            (g_entityDestroyed[flag >> 5] & (1 << (flag & 0x1F))) == expected) {
-            CreateEntityWhenInVerticalRange(D_801D2768);
-        }
-        D_801D2768++;
-    }
-}
-
-void func_801BB140(s16 arg0) {
-    u8 flag;
-    s32 expected;
-
-    if (arg0 < 0) {
-        arg0 = 0;
-    }
-
-    if (D_801D2770 == 0) {
-        func_801BAFEC(arg0 - D_80097908);
-        D_801D2770 = 1;
-    }
-
-    while (true) {
-        if ((D_801D2768->posX == 0xFFFE) || (arg0 > D_801D2768->posX)) {
-            return;
-        }
-
-        expected = 0;
-        flag = (D_801D2768->objectRoomIndex >> 8) + 0xFF;
-        if ((flag == 0xFF) ||
-            (g_entityDestroyed[flag >> 5] & (1 << (flag & 0x1F))) == expected) {
-            CreateEntityWhenInVerticalRange(D_801D2768);
-        }
-        D_801D2768--;
-    }
-}
-
-void func_801BB254(s16 arg0) {
-    while (true) {
-        if ((D_801D276C->posY != (u16)~1) && !(D_801D276C->posY < arg0)) {
-            break;
-        }
-        D_801D276C++;
-    }
-}
-
-void func_801BB2A0(s16 arg0) {
-    while (true) {
-        if ((D_801D276C->posY != 0xFFFF) &&
-            ((arg0 >= D_801D276C->posY) || (D_801D276C->posY == 0xFFFE))) {
-            break;
-        }
-        D_801D276C--;
-    }
-}
-
-void func_801BB2F8(s16 arg0) {
-    u8 flag;
-    s32 expected;
-
-    if (D_801D2774 != 0) {
-        func_801BB254(arg0 - D_8009790C);
-        D_801D2774 = 0;
-    }
-
-    while (true) {
-        if ((D_801D276C->posY == 0xFFFF) || (arg0 < D_801D276C->posY)) {
-            return;
-        }
-
-        expected = 0;
-        flag = (D_801D276C->objectRoomIndex >> 8) + 0xFF;
-        if ((flag == 0xFF) ||
-            (g_entityDestroyed[flag >> 5] & (1 << (flag & 0x1F))) == expected) {
-            CreateEntityWhenInHorizontalRange(D_801D276C);
-        }
-        D_801D276C++;
-    }
-}
-
-void func_801BB3F4(s16 arg0) {
-    u8 flag;
-    s32 expected;
-
-    if (arg0 < 0) {
-        arg0 = 0;
-    }
-
-    if (D_801D2774 == 0) {
-        func_801BB2A0(arg0 - D_8009790C);
-        D_801D2774 = 1;
-    }
-
-    while (true) {
-        if ((D_801D276C->posY == 0xFFFE) || (arg0 > D_801D276C->posY)) {
-            return;
-        }
-
-        expected = 0;
-        flag = (D_801D276C->objectRoomIndex >> 8) + 0xFF;
-        if ((flag == 0xFF) ||
-            (g_entityDestroyed[flag >> 5] & (1 << (flag & 0x1F))) == expected) {
-            CreateEntityWhenInHorizontalRange(D_801D276C);
-        }
-        D_801D276C--;
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", InitRoomEntities);
-
-void func_801BB680(void) {
-    Unkstruct8* currentRoomTileLayout = &g_CurrentRoomTileLayout;
-
-    if (D_80097908 != 0) {
-        s16 tmp = g_Camera.posX.i.hi;
-        if (D_80097908 > 0)
-            func_801BB044(tmp + 320);
-        else
-            func_801BB140(tmp - 64);
-    }
-
-    if (D_8009790C != 0) {
-        s16 tmp = currentRoomTileLayout->unkE;
-        if (D_8009790C > 0)
-            func_801BB2F8(currentRoomTileLayout->unkE + 288);
-        else
-            func_801BB3F4(tmp - 64);
-    }
-}
-
-void CreateEntityFromCurrentEntity(u16 objectId, Entity* entity) {
-    DestroyEntity(entity);
-    entity->objectId = objectId;
-    entity->pfnUpdate = PfnEntityUpdates[objectId];
-    entity->posX.i.hi = g_CurrentEntity->posX.i.hi;
-    entity->posY.i.hi = g_CurrentEntity->posY.i.hi;
-}
-
-void CreateEntityFromEntity(u16 objectId, Entity* source, Entity* entity) {
-    DestroyEntity(entity);
-    entity->objectId = objectId;
-    entity->pfnUpdate = PfnEntityUpdates[objectId];
-    entity->posX.i.hi = source->posX.i.hi;
-    entity->posY.i.hi = source->posY.i.hi;
-}
-
-s32 func_801BB824(Entity* e) {
-    s16 diff;
-
-    diff = PLAYER.posX.i.hi - e->posX.i.hi;
-    diff = ABS(diff);
-
-    if (diff >= 17) {
-        diff = 0;
-    } else {
-        diff = PLAYER.posY.i.hi - e->posY.i.hi;
-        diff = ABS(diff);
-        diff = diff < 33;
-    }
-
-    return diff;
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", EntityRedDoor);
-
-void DestroyEntity(Entity* item) {
-    s32 i;
-    s32 length;
-    u32* ptr;
-
-    if (item->flags & FLAG_FREE_POLYGONS) {
-        g_api.FreePrimitives(item->firstPolygonIndex);
-    }
-
-    ptr = (u32*)item;
-    length = sizeof(Entity) / sizeof(s32);
-    for (i = 0; i < length; i++)
-        *ptr++ = 0;
-}
-
-void DestroyEntityFromIndex(s16 index) {
-    Entity* entity = &g_Entities[index];
-
-    while (entity < &D_8007EF1C) {
-        DestroyEntity(entity);
-        entity++;
-    }
-}
-
-void PreventEntityFromRespawning(Entity* entity) {
-    if (entity->objectRoomIndex) {
-        u32 value = (entity->objectRoomIndex - 1);
-        u16 index = value / 32;
-        u16 bit = value % 32;
-        g_entityDestroyed[index] |= 1 << bit;
-    }
-}
-
-#include "st/AnimateEntity.h"
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801BC6BC);
-
-/*
- * Returns the absolute distance from g_CurrentEntity to player in the X Axis
- */
-s32 GetPlayerDistanceX(void) {
-    s16 yDistance = g_CurrentEntity->posX.i.hi - PLAYER.posX.i.hi;
-
-    if (yDistance < 0) {
-        yDistance = -yDistance;
-    }
-    return yDistance;
-}
-
-/*
- * Returns the absolute distance from g_CurrentEntity to player in the Y Axis
- */
-s32 GetPlayerDistanceY(void) {
-    s32 yDistance = g_CurrentEntity->posY.i.hi - PLAYER.posY.i.hi;
-
-    if (yDistance < 0) {
-        yDistance = -yDistance;
-    }
-    return yDistance;
-}
-
-/**
- * Returns the player's side position relative to g_CurrentEntity
- * 0 = Player is on the right side
- * 1 = Player is on the left side
- * 2 = Player is above
- */
-s16 GetPlayerSide(void) {
-    s16 side = g_CurrentEntity->posX.i.hi > PLAYER.posX.i.hi;
-
-    if (g_CurrentEntity->posY.i.hi > PLAYER.posY.i.hi) {
-        side |= 2;
-    }
-    return side;
-}
-
-void MoveEntity(void) {
-    g_CurrentEntity->posX.val += g_CurrentEntity->accelerationX;
-    g_CurrentEntity->posY.val += g_CurrentEntity->accelerationY;
-}
-
-void FallEntity(void) {
-    if (g_CurrentEntity->accelerationY < FALL_TERMINAL_VELOCITY) {
-        g_CurrentEntity->accelerationY += FALL_GRAVITY;
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801BC8E4);
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801BCB5C);
-
-Entity* AllocEntity(Entity* start, Entity* end) {
-    Entity* current = start;
-
-    while (current < end) {
-        if (current->objectId == 0) {
-            DestroyEntity(current);
-            return current;
-        }
-        current++;
-    }
-    return NULL;
-}
-
-s32 func_801BCDA4(u8 arg0, s16 arg1) { return D_80181A50[arg0] * arg1; }
-
-s16 func_801BCDD0(u8 arg0) { return D_80181A50[arg0]; }
-
-void func_801BCDEC(s32 arg0, s16 arg1) {
-    g_CurrentEntity->accelerationX = func_801BCDA4(arg0, arg1);
-    g_CurrentEntity->accelerationY = func_801BCDA4(arg0 - 0x40, arg1);
-}
-
-u8 func_801BCE58(s16 x, s16 y) { return ((ratan2(y, x) >> 4) + 0x40); }
-
-u8 func_801BCE90(Entity* a, Entity* b) {
-    s32 diffX = (u16)b->posX.i.hi - (u16)a->posX.i.hi;
-    s32 diffY = (u16)b->posY.i.hi - (u16)a->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
-}
-
-u8 func_801BCED8(s32 x, s32 y) {
-    s32 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
-    s32 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return func_801BCE58(diffX, diffY);
-}
-
-u8 func_801BCF20(u8 arg0, u8 arg1, u8 arg2) {
-    u8 var_v0;
-    s8 temp_a2 = arg2 - arg1;
-
-    if (temp_a2 < 0) {
-        var_v0 = -temp_a2;
-    } else {
-        var_v0 = temp_a2;
-    }
-
-    if (var_v0 > arg0) {
-        if (temp_a2 < 0) {
-            var_v0 = arg1 - arg0;
-        } else {
-            var_v0 = arg1 + arg0;
-        }
-
-        return var_v0;
-    }
-
-    return arg2;
-}
-
-void func_801BCF78(u16 slope, s16 speed) {
-    Entity* entity;
-    s32 moveX;
-    s32 moveY;
-
-    moveX = rcos(slope) * speed;
-    entity = g_CurrentEntity;
-    if (moveX < 0) {
-        moveX += 15;
-    }
-    entity->accelerationX = moveX >> 4;
-
-    moveY = rsin(slope) * speed;
-    entity = g_CurrentEntity;
-    if (moveY < 0) {
-        moveY += 15;
-    }
-    entity->accelerationY = moveY >> 4;
-}
-
-u16 func_801BD004(s16 x, s16 y) { return ratan2(y, x); }
-
-u16 func_801BD034(Entity* a, Entity* b) {
-    s32 diffX = b->posX.i.hi - a->posX.i.hi;
-    s32 diffY = b->posY.i.hi - a->posY.i.hi;
-    return ratan2(diffY, diffX);
-}
-
-u16 func_801BD06C(s32 x, s32 y) {
-    s16 diffX = x - (u16)g_CurrentEntity->posX.i.hi;
-    s16 diffY = y - (u16)g_CurrentEntity->posY.i.hi;
-    return ratan2(diffY, diffX);
-}
-
-u16 func_801BD0B4(u16 arg0, s16 arg1, s16 arg2) {
-    u16 var_v0 = arg1;
-    u16 temp_a2 = arg2 - arg1;
-    u16 var_v0_2;
-
-    if (temp_a2 & 0x800) {
-        var_v0_2 = (0x800 - temp_a2) & 0x7FF;
-    } else {
-        var_v0_2 = temp_a2;
-    }
-
-    if (var_v0_2 > arg0) {
-        if (temp_a2 & 0x800) {
-            var_v0 = arg1 - arg0;
-        } else {
-            var_v0 = arg1 + arg0;
-        }
-
-        return var_v0;
-    }
-
-    return arg2;
-}
-
-void func_801BD114(u8 step) {
-    g_CurrentEntity->step = step;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
-}
-
-void func_801BD134(u8 arg0) {
-    g_CurrentEntity->step_s = arg0;
-    g_CurrentEntity->animFrameIdx = 0;
-    g_CurrentEntity->animFrameDuration = 0;
-}
-
-void func_801BD150(u16 arg0, u16 sfxId) {
-    if (sfxId != 0) {
-        func_801C2598(sfxId);
-    }
-
-    if (arg0 == 0xFF) {
-        DestroyEntity(g_CurrentEntity);
-        return;
-    }
-
-    g_CurrentEntity->objectId = ENTITY_EXPLOSION;
-    g_CurrentEntity->pfnUpdate = (PfnEntityUpdate)EntityExplosion;
-    g_CurrentEntity->subId = arg0;
-    g_CurrentEntity->animCurFrame = 0;
-    g_CurrentEntity->unk19 = 0;
-    g_CurrentEntity->step = 0;
-    g_CurrentEntity->step_s = 0;
-}
-
-void InitializeEntity(u16 arg0[]) {
-    u16 enemyId;
-    EnemyDef* enemyDef;
-
-    g_CurrentEntity->animSet = *arg0++;
-    g_CurrentEntity->animCurFrame = *arg0++;
-    g_CurrentEntity->unk5A = *arg0++;
-    g_CurrentEntity->palette = *arg0++;
-
-    enemyId = *arg0++;
-    g_CurrentEntity->enemyId = enemyId;
-    enemyDef = &g_api.enemyDefs[enemyId];
-    g_CurrentEntity->hitPoints = enemyDef->hitPoints;
-    g_CurrentEntity->attack = enemyDef->attack;
-    g_CurrentEntity->attackElement = enemyDef->attackElement;
-    g_CurrentEntity->unk3C = enemyDef->unkC;
-    g_CurrentEntity->hitboxWidth = enemyDef->hitboxWidth;
-    g_CurrentEntity->hitboxHeight = enemyDef->hitboxHeight;
-    g_CurrentEntity->flags = enemyDef->unk24;
-    g_CurrentEntity->unk10 = 0;
-    g_CurrentEntity->unk12 = 0;
-    g_CurrentEntity->step_s = 0;
-    g_CurrentEntity->step++;
-    if (g_CurrentEntity->zPriority == 0) {
-        g_CurrentEntity->zPriority = g_zEntityCenter.S16.unk0 - 0xC;
-    }
-}
-
-void EntityDummy(Entity* arg0) {
-    if (arg0->step == 0) {
-        arg0->step++;
-    }
-}
-
-s32 func_801BD308(u16* hitSensors, s16 sensorCount) {
-    s32 accelerationX = g_CurrentEntity->accelerationX;
-    Collider collider;
-    u16 temp_a1;
-    s16 x;
-    s16 y;
-    s16 i;
-
-    if (accelerationX != 0) {
-        x = g_CurrentEntity->posX.i.hi;
-        y = g_CurrentEntity->posY.i.hi;
-        for (i = 0; i < sensorCount; i++) {
-            if (accelerationX < 0) {
-                s16 newX = x + *hitSensors++;
-                x = newX;
-            } else {
-                s16 newX = x - *hitSensors++;
-                x = newX;
-            }
-
-            y += *hitSensors++;
-            g_api.CheckCollision(x, y, &collider, 0);
-            if (collider.unk0 & 2 &&
-                ((!(collider.unk0 & 0x8000)) || (i != 0))) {
-                return 2;
-            }
-        }
-        return 0;
-    }
-}
-
-void func_801BD430(u16* hitSensors, s16 sensorCount) {
-    s32 accelerationX = g_CurrentEntity->accelerationX;
-    Collider collider;
-    s16 x;
-    s16 y;
-    s16 i;
-
-    if (accelerationX == 0)
-        return;
-
-    x = g_CurrentEntity->posX.i.hi;
-    y = g_CurrentEntity->posY.i.hi;
-
-    for (i = 0; i < sensorCount; i++) {
-        if (accelerationX < 0) {
-            x = x + *hitSensors++;
-        } else {
-            x = x - *hitSensors++;
-        }
-
-        y += *hitSensors++;
-        g_api.CheckCollision(x, y, &collider, 0);
-        if (collider.unk0 & 2 && (!(collider.unk0 & 0x8000) || i != 0)) {
-            if (accelerationX < 0) {
-                g_CurrentEntity->posX.i.hi += collider.unk1C;
-            } else {
-                g_CurrentEntity->posX.i.hi += collider.unk14;
-            }
-            return;
-        }
-    }
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", func_801BD588);
-
-void ReplaceBreakableWithItemDrop(Entity* self) {
-    u16 subId;
-
-    PreventEntityFromRespawning(self);
-
-#if STAGE != STAGE_ST0
-    if (!(g_Status.relics[10] & 2)) {
-        DestroyEntity(self);
-        return;
-    }
-#endif
-
-    subId = self->subId &= 0xFFF;
-
-    if (subId < 0x80) {
-        self->objectId = ENTITY_PRICE_DROP;
-        self->pfnUpdate = (PfnEntityUpdate)EntityPrizeDrop;
-        self->animFrameDuration = 0;
-        self->animFrameIdx = 0;
-    } else {
-        subId -= 0x80;
-        self->objectId = ENTITY_INVENTORY_DROP;
-        self->pfnUpdate = (PfnEntityUpdate)EntityEquipItemDrop;
-    }
-
-    self->subId = subId;
-    self->unk6D = 0x10;
-    self->step = 0;
-}
-
-void func_801BD984(void) {
-    s32 temp_v1;
-    Entity* entity;
-
-    entity = g_CurrentEntity;
-    if (entity->accelerationY >= 0) {
-        temp_v1 =
-            entity->ext.generic.unk88.S16.unk0 + entity->ext.generic.unk84.unk;
-        entity->ext.generic.unk84.unk = temp_v1;
-        entity->accelerationX = temp_v1;
-        if (temp_v1 == 0x10000 || temp_v1 == -0x10000) {
-            entity->ext.generic.unk88.S16.unk0 =
-                -entity->ext.generic.unk88.S16.unk0;
-        }
-        entity = g_CurrentEntity;
-    }
-    NOP;
-
-    if (entity->accelerationY < 0x00004000) {
-        entity->accelerationY += 0x2000;
-    }
-}
-
-void func_801BDA08(u16 arg0) {
-    Collider res;
-
-    if (g_CurrentEntity->accelerationX < 0) {
-        g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                             g_CurrentEntity->posY.i.hi - 7, &res, 0);
-        if (res.unk0 & 5) {
-            g_CurrentEntity->accelerationY = 0;
-        }
-    }
-
-    g_api.CheckCollision(g_CurrentEntity->posX.i.hi,
-                         g_CurrentEntity->posY.i.hi + 7, &res, 0);
-
-    if (arg0) {
-        if (!(res.unk0 & 5)) {
-            MoveEntity();
-            FallEntity();
-            return;
-        }
-
-        g_CurrentEntity->accelerationX = 0;
-        g_CurrentEntity->accelerationY = 0;
-
-        if (res.unk0 & 4) {
-            g_CurrentEntity->posY.val += 0x2000;
-        } else {
-            g_CurrentEntity->posY.i.hi += res.unk18;
-        }
-    } else {
-        if (!(res.unk0 & 5)) {
-            MoveEntity();
-            func_801BD984();
-        }
-    }
-}
-
-void CollectHeart(u16 heartSize) {
-    s32* hearts;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    hearts = &g_Status.hearts;
-    *hearts += c_HeartPrizes[heartSize];
-
-    if (g_Status.heartsMax < *hearts) {
-        *hearts = g_Status.heartsMax;
-    }
-
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectGold(u16 goldSize) {
-    s32 *gold, *unk;
-    u16 goldSizeIndex;
-
-    g_api.PlaySfx(NA_SE_PL_COLLECT_GOLD);
-    gold = &g_Status.gold;
-    goldSizeIndex = goldSize - 2;
-    *gold += c_GoldPrizes[goldSizeIndex];
-    if (*gold > MAX_GOLD) {
-        *gold = MAX_GOLD;
-    }
-
-    unk = &D_80097410;
-    if (*unk) {
-        g_api.FreePrimitives(D_80097414);
-        *unk = 0;
-    }
-
-    func_801C3E14(D_80181DC4[goldSizeIndex], 1);
-    DestroyEntity(g_CurrentEntity);
-}
-
-INCLUDE_ASM("asm/us/st/np3/nonmatchings/3246C", CollectSubweapon);
-
-void CollectHeartVessel(void) {
-    if (g_CurrentPlayableCharacter != PLAYER_ALUCARD) {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_Status.hearts += HEART_VESSEL_RICHTER;
-
-        if (g_Status.heartsMax < g_Status.hearts) {
-            g_Status.hearts = g_Status.heartsMax;
-        }
-    } else {
-        g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-        g_api.func_800FE044(HEART_VESSEL_INCREASE, 0x4000);
-    }
-    DestroyEntity(g_CurrentEntity);
-}
-
-void CollectLifeVessel(void) {
-    g_api.PlaySfx(NA_SE_PL_COLLECT_HEART);
-    g_api.func_800FE044(5, 0x8000);
-    DestroyEntity(g_CurrentEntity);
-}
-
-void DestroyCurrentEntity(void) { DestroyEntity(g_CurrentEntity); }
